Fix a bug that stuck slash command from working
If we already in combat when we login, InCombat variable would default to
false, and the game won't fire PLAYER_REGEN_DISABLED event till we leave
that combat. And UnitAffectingCombat("player") also not work during
login.

This would stuck slash command to work because it would check InCombat.

Added a delayed in-combat check when login.